requirements.txt
pydantic
uvicorn
psycopg2
paypayopa
fastapi
python-dotenv
SQLAlchemy
Pillow
firebase-admin
pymysql
